Product_query/requirements.txt
pydantic>=2.7.0,<3.0.0
streamlit>=1.39.0
groq>=0.9.0
python-dotenv>=1.0.1